% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{g_index}
\alias{g_index}
\title{g index}
\usage{
g_index(citations)
}
\arguments{
\item{citations}{A numeric vector containing citations}
}
\value{
return the g index for the given citations
}
\description{
g index
}
\examples{
g_index(c(1,2,5,0,3,11))
}
\seealso{
\code{\link[=h_index]{h index}}
}

