% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{h_index}
\alias{h_index}
\title{h index}
\usage{
h_index(citations)
}
\arguments{
\item{citations}{A numeric vector containing citations}
}
\value{
return the h index for the given citations
}
\description{
Find h index for a given set of documents
}
\examples{
h_index(c(1,2,5,0,3,11))
}
\references{
Hirsch, J. E. (2005). An index to quantify an individual's
scientific research output. Proceedings of the National academy of Sciences
 of the United States of America, 102(46), 16569-16572.
}
\seealso{
\code{\link{g_index}}
}

