% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textM.R
\name{term_freq}
\alias{term_freq}
\title{Term Frequency}
\usage{
term_freq(text, pub_years = NULL, sep = NULL, top = NULL)
}
\arguments{
\item{text}{A character vector}

\item{pub_years}{A numeric vector containing publication years}

\item{sep}{A character value whcih separates the terms (optional)}

\item{top}{The number of terms to return}
}
\value{
Term frequencey vector or matrix (for year-wise)
}
\description{
Term Frequency
}

